Backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.4.2
requests==2.31.0
pandas==2.1.1
scikit-learn==1.3.2
sqlalchemy==2.0.22
python-dotenv==1.0.0
pytest==7.4.3
matplotlib==3.8.0
seaborn==0.13.0
